khp: Track allocation locality and optimize freeing
Use one byte in the khp_meta struct to track whether the allocation has so
far only been used in a CPU-local manner, and if so, on which CPU.

The idea here is that many workloads will rapidly allocate and free memory
for the current task, e.g. for temporary allocations in syscall context or
for resources that userspace quickly opens and closes (e.g. struct file).
In those cases, it would be unfortunate to always have to wait for all
other cores to be done with the allocations, with all the synchronization
that entails; instead, we can directly release the allocations on syscall
exit.

However, there are still workloads that will create many temporary
allocations that will be marked as global, e.g. because of inter-process
communication over unix domain sockets. (We could try to add
infrastructure to let the producer "relinquish" the allocations, but that
would probably become significantly more ugly than this patch.)

A straightforward implementation would be to e.g. use 0x00 or 0xff for
"global allocation" and use all other numbers to identify CPU cores.
However, a disadvantage of that scheme is that the test in the
pointer-decoding fastpath would have to be written using two separate
comparisons - and we really care about every instruction in the fastpath.

Instead, we give each CPU an ID with hamming weight 4 (exactly half the
bits are set), store the inverted ID in the khp_meta struct, and replace
the inverted ID with 0 when an allocation is no longer CPU-local.

This allows us to write the check in __khp_decode_ptr as a logical AND
between the expected ID and the stored ID: If the object is zero, the
result is zero (? & 0 == 0); if the object is local to the current CPU, the
result is also zero (~ID & ID == 0); but if the object belongs to a remote
CPU, the result is nonzero (because when two values with the same hamming
weight are different, each value must have at least one bit set that is
unset in the other one).