khp: Fast, specialized handling for single loads
The relative cost of __khp_decode_ptr() is worst when the decoded pointer
is only used for a single load which is not in a loop below the pointer
source.

__khp_decode_ptr() is costly because it has to track that the current task
is referencing the provided pointer; in particular, it has to:

 - handle userspace pointers just like raw kernel pointers (in case we're
   called from userspace accessors)
 - check whether a local allocation is being accessed from a different CPU
 - write the metadata pointer into the associated stack slot

But we don't really need any of this to perform a single load with UAF
protection: We can just perform the load first, then do the cookie check
afterwards (and on X86, we don't have to worry about load-load reordering).

Luckily a load alone should have no overly terrible effects as long as the
loaded value can't actually be used until we know that the allocation was
still live at the time of load.

The worst that could happen with this would be memory-mapped I/O, if it
is possible for MMIO addresses to appear in the metadata structure - that
is probably possible with the current implementation.
It's probably worth revisiting whether X86 has any device where wild MMIO
reads can actually cause dangerous state changes, and if so, we'll have to
harden the use of the first half of `struct khp_meta` to avoid ever having
pointer-like values in there that refer to MMIO memory.
(If that is a problem, anything that uses probe_kernel_read() on pointers
that are not guaranteed to be pointers into normal RAM should probably also
be revisited.)

Signed-off-by: Jann Horn <jannh@google.com>