mm/slub: Refactor page freelist+counters into named struct
[[[ abandoned in-progress work, not ready for inclusion ]]]

Refactor SLUB's freelist and counters together into a separate named struct.
The main motivation here is that once a following commit shrinks that
struct to a single word, we'll be able to cleanly use the cmpxchg()
helper to operate on it.
As a nice side effect, this also gets rid of stack variables of type "struct
page".

Signed-off-by: Jann Horn <jannh@google.com>