mm/slub: Optionally permit reducing alignment of struct page
[[[ abandoned in-progress work, not ready for inclusion ]]]

On 64-bit kernels without CONFIG_MEMCG, struct page currently has an
8-byte hole at the end. With PAGE_SIZE==4096, that ends up costing us
about 0.2% (8/4096) of available memory.

This hole is caused by SLUB, which wants to use cmpxchg_double to
atomically update two words in struct page that consist of a freelist
head pointer, a bitfield with 32 bits, and 4 bytes of padding.
SLUB can also function without cmpxchg_double, but in that case it has
to fall back to using a bitlock to protect the page, which means
multiple atomic operations and potentially expensive contention, so it'd
probably be good to avoid that.

To reduce the padding of struct page without giving up on lockless
updates, store only the lower 32 bits of the freelist head pointer;
since the page allocator guarantees that order-N pages have order-N
alignment, slab pages can't straddle 4 GiB boundaries, so we can
reconstruct the full freelist head pointer using either page_to_virt()
or any pointer into the page.

Place the truncated pointer in the more significant half of the 8-byte
word so that the value used for cmpxchg could be computed as
"full_pointer << 32) | counters", in case that helps the compiler
optimize things.

An alternative to this approach would be to let pages alternatingly have
the "slab_cache" member before or after the freelist+counters, with some
gross address-based rounding to calculate the addresses of these three
members. That seems to be like it would be uglier though.

According to Agner Fog's tables, on Skylake, LOCK CMPXCHG16B costs
25 uops and has a reciprocal throughput of 26 cycles, while LOCK CMPXCHG
only costs 10 uops with a reciprocal throughput of 18 cycles; so only
doing a word-sized cmpxchg might also compensate for the overhead of the
extra arithmetic.

Signed-off-by: Jann Horn <jannh@google.com>