mm/slub: Get rid of slab_free_freelist_hook()
Currently, the SLUB bulk freeing path works as follows:

 - kmem_cache_free_bulk() receives an array of objects
 - build_detached_freelist() turns the array into per-slab detached
   freelists
 - slab_free_freelist_hook() iterates through the detached freelist and:
    - calls slab_free_hook() on the objects (possibly handing some of them
      off to KASAN)
    - rewrites the freelist (in case some objects were handed off to KASAN)

slab_free_freelist_hook() also writes a NULL freelist entry into the last
element of the detached freelist.

Because most of the object-freeing path is shared between the
single-element and bulk entry points for freeing objects, that terminating
NULL freelist entry is unnecessarily written (and encoded, if
CONFIG_SLAB_FREELIST_HARDENED is enabled) even in the compiled code for
the single-element entry point kmem_cache_free(), and even if the kernel
was not built with KASAN support.

In theory, this also means that we're unnecessarily calling
get_freepointer() on memory that does not actually contain a freelist
element, but that has no effect in the single-element-freeing case and is
compiled out, so it currently doesn't matter.
However, it would cause problems if we wanted to add validation in
get_freepointer().

Move the slab_free_hook() calls before the point where a freelist has been
assembled. This lets us get rid of the ugly slab_free_freelist_hook() and
saves some instructions in kmem_cache_free() in typical builds.

This might also make bulk freeing more efficient.

Here's the relevant portion of the diff between kmem_cache_free() before
and after this patch (from a build without CONFIG_MEMCG and with
CONFIG_SLAB_FREELIST_HARDENED; offsets replaced with "???" to reduce diff
noise):

        add    rax,QWORD PTR [rip+0x???]
                        R_X86_64_PC32   vmemmap_base-0x4
        mov    rdx,QWORD PTR [rax+0x???]
-       mov    r12,rax
+       mov    rbp,rax
        test   dl,0x1
        jne    <kmem_cache_free+0x???>
        nop    DWORD PTR [rax+rax*1+0x???]
-       mov    rax,QWORD PTR [r12]
+       mov    rax,QWORD PTR [rbp+0x???]
        test   ah,0x2
        mov    eax,0x0
-       cmove  r12,rax
+       cmove  rbp,rax
        nop    DWORD PTR [rax+rax*1+0x???]
-       mov    eax,DWORD PTR [rbp+0x???]
-       add    rax,rbx
-       test   rbx,rbx
-       je     <kmem_cache_free+0x???>
-       mov    rdx,rax
-       bswap  rdx
-       xor    rdx,QWORD PTR [rbp+0x???]
-       mov    QWORD PTR [rax],rdx
-       mov    rax,QWORD PTR [rbp+0x???]
+       mov    rax,QWORD PTR [r14]
        add    rax,QWORD PTR gs:[rip+0x???]
                        R_X86_64_PC32   this_cpu_off-0x4
        mov    rdx,QWORD PTR [rax+0x???]
-       cmp    QWORD PTR [rax+0x???],r12
+       cmp    QWORD PTR [rax+0x???],rbp
        jne    <kmem_cache_free+0x???>
-       mov    ecx,DWORD PTR [rbp+0x???]
+       mov    ecx,DWORD PTR [r14+0x???]
        mov    rax,QWORD PTR [rax]
        add    rcx,rbx
        cmp    rbx,rax
        je     <kmem_cache_free+0x???>
-       mov    rsi,QWORD PTR [rbp+0x???]
+       mov    rsi,QWORD PTR [r14+0x???]
        mov    rdi,rcx
        bswap  rdi
        xor    rsi,rax
        xor    rsi,rdi
        mov    QWORD PTR [rcx],rsi
        lea    rcx,[rdx+0x???]
-       mov    rdi,QWORD PTR [rbp+0x???]
+       mov    rdi,QWORD PTR [r14]
        lea    rsi,[rdi]
        call   <kmem_cache_free+0x???>
                        R_X86_64_PLT32  this_cpu_cmpxchg16b_emu-0x4

As you can see, this gets rid of an encoded freelist element storing
sequence (add s->offset to pointer, bswap, xor, store).

TODO: microbenchmark or something

Signed-off-by: Jann Horn <jannh@google.com>